+++ compendium/DeclarativeServices/src/manager/ReferenceManagerImpl.cpp
@@ -20,6 +20,8 @@
 
   =============================================================================*/
 #include <cassert>
+#include <memory>
+
 #include "cppmicroservices/ServiceReference.h"
 #include "cppmicroservices/LDAPProp.h"
 #include "cppmicroservices/servicecomponent/ComponentConstants.hpp"
@@ -55,14 +57,28 @@ LDAPFilter GetReferenceLDAPFilter(const metadata::ReferenceMetadata& refMetadata
   return LDAPFilter(expr);
 }
 
-ReferenceManagerImpl::ReferenceManagerImpl(const metadata::ReferenceMetadata& metadata,
-                                           const cppmicroservices::BundleContext& bc,
-                                           std::shared_ptr<cppmicroservices::logservice::LogService> logger,
-                                           const std::string& configName)
+ReferenceManagerBaseImpl::ReferenceManagerBaseImpl(const metadata::ReferenceMetadata& metadata
+                                           , const cppmicroservices::BundleContext& bc
+                                           , std::shared_ptr<cppmicroservices::logservice::LogService> logger
+                                           , const std::string& configName)
+  : ReferenceManagerBaseImpl(metadata
+                             , bc
+                             , logger
+                             , configName
+                             , CreateBindingPolicy(*this, metadata.policy, metadata.policyOption))
+{
+}
+
+ReferenceManagerBaseImpl::ReferenceManagerBaseImpl(const metadata::ReferenceMetadata& metadata
+                                           , const cppmicroservices::BundleContext& bc
+                                           , std::shared_ptr<cppmicroservices::logservice::LogService> logger
+                                           , const std::string& configName
+                                           , std::unique_ptr<BindingPolicy> policy)
   : metadata(metadata)
   , tracker(nullptr)
   , logger(std::move(logger))
   , configName(configName)
+  , bindingPolicy(std::move(policy))
 {
   if(!bc || !this->logger)
   {
@@ -81,7 +97,7 @@ ReferenceManagerImpl::ReferenceManagerImpl(const metadata::ReferenceMetadata& me
   }
 }
 
-void ReferenceManagerImpl::StopTracking()
+void ReferenceManagerBaseImpl::StopTracking()
 {
   try
   {
@@ -93,36 +109,29 @@ void ReferenceManagerImpl::StopTracking()
   }
 }
 
-std::set<cppmicroservices::ServiceReferenceBase> ReferenceManagerImpl::GetBoundReferences() const
+std::set<cppmicroservices::ServiceReferenceBase> ReferenceManagerBaseImpl::GetBoundReferences() const
 {
   auto boundRefsHandle = boundRefs.lock();
   return std::set<cppmicroservices::ServiceReferenceBase>(boundRefsHandle->begin(), boundRefsHandle->end());
 }
 
-std::set<cppmicroservices::ServiceReferenceBase> ReferenceManagerImpl::GetTargetReferences() const
+std::set<cppmicroservices::ServiceReferenceBase> ReferenceManagerBaseImpl::GetTargetReferences() const
 {
   auto matchedRefsHandle = matchedRefs.lock();
   return std::set<cppmicroservices::ServiceReferenceBase>(matchedRefsHandle->begin(), matchedRefsHandle->end());
 }
 
-// util method to extract service-id from a given reference
-long GetServiceId(const ServiceReferenceBase& sRef)
-{
-  auto idAny = sRef.GetProperty(cppmicroservices::Constants::SERVICE_ID);
-  return cppmicroservices::any_cast<long>(idAny);
-}
-
-bool ReferenceManagerImpl::IsOptional() const
+bool ReferenceManagerBaseImpl::IsOptional() const
 {
   return (metadata.minCardinality == 0);
 }
 
-bool ReferenceManagerImpl::IsSatisfied() const
+bool ReferenceManagerBaseImpl::IsSatisfied() const
 {
   return (boundRefs.lock()->size() >= metadata.minCardinality);
 }
 
-ReferenceManagerImpl::~ReferenceManagerImpl()
+ReferenceManagerBaseImpl::~ReferenceManagerBaseImpl()
 {
   StopTracking();
 }
@@ -130,13 +139,14 @@ ReferenceManagerImpl::~ReferenceManagerImpl()
 struct dummyRefObj {
 };
 
-bool ReferenceManagerImpl::UpdateBoundRefs()
+bool ReferenceManagerBaseImpl::UpdateBoundRefs()
 {
   auto matchedRefsHandle = matchedRefs.lock(); // acquires lock on matchedRefs
   const auto matchedRefsHandleSize = matchedRefsHandle->size();
   if(matchedRefsHandleSize >= metadata.minCardinality)
   {
     auto boundRefsHandle = boundRefs.lock(); // acquires lock on boundRefs
+    boundRefsHandle->clear();
     std::copy_n(matchedRefsHandle->rbegin(),
                 std::min(metadata.maxCardinality, matchedRefsHandleSize),
                 std::inserter(*(boundRefsHandle),
@@ -147,91 +157,7 @@ bool ReferenceManagerImpl::UpdateBoundRefs()
   // release locks on matchedRefs and boundRefs
 }
 
-// This method implements the following algorithm
-//
-//  if reference becomes satisfied
-//    Copy service references from #matchedRefs to #boundRefs
-//    send a SATISFIED notification to listeners
-//  else if reference is already satisfied
-//    if policyOption is reluctant
-//      ignore the new servcie
-//    else if policyOption is GREEDY
-//      if the new service is better than any of the existing services in #boundRefs
-//        send UNSATISFIED notification to listeners
-//        clear #boundRefs
-//        copy #matchedRefs to #boundRefs
-//        send a SATISFIED notification to listeners
-//      endif
-//    endif
-//  endif
-void ReferenceManagerImpl::ServiceAdded(const cppmicroservices::ServiceReferenceBase& reference)
-{
-  std::vector<RefChangeNotification> notifications;
-  if(!reference)
-  {
-    logger->Log(SeverityLevel::LOG_DEBUG, "ServiceAdded: service with id " + std::to_string(GetServiceId(reference)) + " has already been unregistered, no-op");
-    return;
-  }
-  // const auto minCardinality = metadata.minCardinality;
-  // const auto maxCardinality = metadata.maxCardinality;
-  // auto prevSatisfied = false;
-  // auto becomesSatisfied = false;
-  auto replacementNeeded = false;
-  auto notifySatisfied = false;
-  auto serviceIdToUnbind = -1;
-
-  if(!IsSatisfied())
-  {
-    notifySatisfied = UpdateBoundRefs(); // becomes satisfied if return value is true
-  }
-  else // previously satisfied
-  {
-    if (metadata.policyOption == "greedy")
-    {
-      auto boundRefsHandle = boundRefs.lock(); // acquire lock on boundRefs
-      if (boundRefsHandle->find(reference) == boundRefsHandle->end()) // reference is not bound yet
-      {
-        if (!boundRefsHandle->empty())
-        {
-          const ServiceReferenceBase& minBound = *(boundRefsHandle->begin());
-          if (minBound < reference)
-          {
-            replacementNeeded = true;
-            serviceIdToUnbind = GetServiceId(minBound);
-          }
-        }
-        else
-        {
-          replacementNeeded = IsOptional();
-        }
-      }
-    }
-  }
-
-  if(replacementNeeded)
-  {
-    logger->Log(SeverityLevel::LOG_DEBUG, "Notify UNSATISFIED for reference " + metadata.name);
-    RefChangeNotification notification{metadata.name, RefEvent::BECAME_UNSATISFIED};
-    notifications.push_back(std::move(notification));
-    // The following "clear and copy" strategy is sufficient for
-    // updating the boundRefs for static binding policy
-    if(0 < serviceIdToUnbind)
-    {
-      auto boundRefsHandle = boundRefs.lock();
-      boundRefsHandle->clear();
-    }
-    notifySatisfied = UpdateBoundRefs();
-  }
-  if(notifySatisfied)
-  {
-    logger->Log(SeverityLevel::LOG_DEBUG, "Notify SATISFIED for reference " + metadata.name);
-    RefChangeNotification notification{metadata.name, RefEvent::BECAME_SATISFIED};
-    notifications.push_back(std::move(notification));
-  }
-  BatchNotifyAllListeners(notifications);
-}
-
-cppmicroservices::InterfaceMapConstPtr ReferenceManagerImpl::AddingService(const cppmicroservices::ServiceReference<void>& reference)
+cppmicroservices::InterfaceMapConstPtr ReferenceManagerBaseImpl::AddingService(const cppmicroservices::ServiceReference<void>& reference)
 {
   // Each service registered by DS contains a service property representing the component configuration name
   // to which it belongs. By checking the component configuration name of a service it can be determined
@@ -241,8 +167,9 @@ cppmicroservices::InterfaceMapConstPtr ReferenceManagerImpl::AddingService(const
   // ASSUMPTION: If there is no component configuration name then its assumed this service was not registered by
   // DS and could not satisfy itself since it is not managed by DS.
   auto const compConfigName = reference.GetProperty(COMPONENT_NAME);
-  if ((!compConfigName.Empty() && configName != compConfigName.ToStringNoExcept()) ||
-    compConfigName.Empty()) { 
+  if ((true == compConfigName.Empty())
+      || (configName != compConfigName.ToStringNoExcept()))
+  { 
     // acquire lock on matchedRefs
     auto matchedRefsHandle = matchedRefs.lock();
     matchedRefsHandle->insert(reference);
@@ -253,60 +180,17 @@ cppmicroservices::InterfaceMapConstPtr ReferenceManagerImpl::AddingService(const
   // https://osgi.org/download/r6/osgi.core-6.0.0.pdf#page=432. Sporadically not returning a valid service
   // when a user calls getService, due to a service's references still resolving, was deemed undesirable
   // for user workflows.
-  ServiceAdded(reference);
+  bindingPolicy->ServiceAdded(reference);
 
   // A non-null object must be returned to indicate to the ServiceTracker that
   // we are tracking the service and need to be called back when the service is removed.
   return MakeInterfaceMap<dummyRefObj>(std::make_shared<dummyRefObj>());
 }
 
-void ReferenceManagerImpl::ModifiedService(const cppmicroservices::ServiceReference<void>& /*reference*/,
+void ReferenceManagerBaseImpl::ModifiedService(const cppmicroservices::ServiceReference<void>& /*reference*/,
                                            const cppmicroservices::InterfaceMapConstPtr& /*service*/)
 {
   // no-op since there is no use case for property update
-}
-
-/**
- *This method implements the following algorithm
- *
- * If the removed service is found in the #boundRefs
- *   send a UNSATISFIED notification to listeners
- *   clear the #boundRefs member
- *   copy #matchedRefs to #boundRefs
- *   if reference is still satisfied
- *     send a SATISFIED notification to listeners
- *  endif
- * endif
- */
-void ReferenceManagerImpl::ServiceRemoved(const cppmicroservices::ServiceReferenceBase& reference)
-{
-  auto removeBoundRef = false;
-  std::vector<RefChangeNotification> notifications;
-
-  { // acquire lock on boundRefs
-    auto boundRefsHandle = boundRefs.lock();
-    auto itr = boundRefsHandle->find(reference);
-    removeBoundRef = (itr != boundRefsHandle->end());
-  } // end lock on boundRefs
-
-  if(removeBoundRef)
-  {
-    logger->Log(SeverityLevel::LOG_DEBUG, "Notify UNSATISFIED for reference " + metadata.name);
-    RefChangeNotification notification { metadata.name, RefEvent::BECAME_UNSATISFIED };
-    notifications.push_back(std::move(notification));
-    {
-      auto boundRefsHandle = boundRefs.lock();
-      boundRefsHandle->clear();
-    }
-    auto notifySatisfied = UpdateBoundRefs();
-    if(notifySatisfied)
-    {
-      logger->Log(SeverityLevel::LOG_DEBUG, "Notify SATISFIED for reference " + metadata.name);
-      RefChangeNotification notification{metadata.name, RefEvent::BECAME_SATISFIED};
-      notifications.push_back(std::move(notification));
-    }
-    BatchNotifyAllListeners(notifications);
-  }
 }
 
 /**
@@ -314,7 +198,7 @@ void ReferenceManagerImpl::ServiceRemoved(const cppmicroservices::ServiceReferen
  * the component configuration must be reactivated and the replacement service is bound to
  * the new component instance.
  */
-void ReferenceManagerImpl::RemovedService(const cppmicroservices::ServiceReference<void>& reference,
+void ReferenceManagerBaseImpl::RemovedService(const cppmicroservices::ServiceReference<void>& reference,
                                           const cppmicroservices::InterfaceMapConstPtr& /*service*/)
 {
   { // acquire lock on matchedRefs
@@ -326,15 +210,15 @@ void ReferenceManagerImpl::RemovedService(const cppmicroservices::ServiceReferen
   // This behavior deviates from what is described in the "synchronous" section in
   // https://osgi.org/download/r6/osgi.core-6.0.0.pdf#page=432. Sometimes not returning a valid service
   // due to a service's references still resolving was deemed undesirable for user workflows.
-  ServiceRemoved(reference);
+  bindingPolicy->ServiceRemoved(reference);
 }
 
-std::atomic<cppmicroservices::ListenerTokenId> ReferenceManagerImpl::tokenCounter(0);
+std::atomic<cppmicroservices::ListenerTokenId> ReferenceManagerBaseImpl::tokenCounter(0);
 
 /**
  * Method is used to register a listener for callbacks
  */
-cppmicroservices::ListenerTokenId ReferenceManagerImpl::RegisterListener(std::function<void(const RefChangeNotification&)> notify)
+cppmicroservices::ListenerTokenId ReferenceManagerBaseImpl::RegisterListener(std::function<void(const RefChangeNotification&)> notify)
 {
   auto notifySatisfied = UpdateBoundRefs();
   if(notifySatisfied)
@@ -354,7 +238,7 @@ cppmicroservices::ListenerTokenId ReferenceManagerImpl::RegisterListener(std::fu
 /**
  * Method is used to remove a registered listener
  */
-void ReferenceManagerImpl::UnregisterListener(cppmicroservices::ListenerTokenId token)
+void ReferenceManagerBaseImpl::UnregisterListener(cppmicroservices::ListenerTokenId token)
 {
   auto listenerMapHandle = listenersMap.lock();
   listenerMapHandle->erase(token);
@@ -363,7 +247,7 @@ void ReferenceManagerImpl::UnregisterListener(cppmicroservices::ListenerTokenId 
 /**
  * Method used to notify all listeners
  */
-void ReferenceManagerImpl::BatchNotifyAllListeners(const std::vector<RefChangeNotification>& notifications) noexcept
+void ReferenceManagerBaseImpl::BatchNotifyAllListeners(const std::vector<RefChangeNotification>& notifications) noexcept
 {
   if (notifications.empty() || listenersMap.lock()->empty())
   {
@@ -384,5 +268,35 @@ void ReferenceManagerImpl::BatchNotifyAllListeners(const std::vector<RefChangeNo
     }
   }
 }
+
+// util method to extract service-id from a given reference
+long ReferenceManagerBaseImpl::GetServiceId(const ServiceReferenceBase& sRef)
+{
+  auto idAny = sRef.GetProperty(cppmicroservices::Constants::SERVICE_ID);
+  return cppmicroservices::any_cast<long>(idAny);
+}
+
+std::unique_ptr<ReferenceManagerBaseImpl::BindingPolicy> ReferenceManagerBaseImpl::CreateBindingPolicy(ReferenceManagerBaseImpl& ref
+                                                                                                       , const std::string& policy
+                                                                                                       , const std::string& policyOption)
+{
+  if (policy == "static") {
+    if (policyOption == "reluctant") {
+      return std::make_unique<BindingPolicyStaticReluctant>(ref);
+    }
+    else { // greedy
+      return std::make_unique<BindingPolicyStaticGreedy>(ref);
+    }
+  }
+  else { // dynamic
+    if (policyOption == "reluctant") {
+      return std::make_unique<BindingPolicyDynamicReluctant>(ref);
+    }
+    else { // greedy
+      return std::make_unique<BindingPolicyDynamicGreedy>(ref);
+    } 
+  }
+}
+
 }
 }