+++ compendium/DeclarativeServices/src/manager/states/CMDisabledState.hpp
@@ -23,7 +23,11 @@
 #ifndef CMDisabledState_hpp
 #define CMDisabledState_hpp
 
+#if defined(USING_GTEST)
 #include "gtest/gtest_prod.h"
+#else
+#define FRIEND_TEST(x, y)
+#endif
 #include "ComponentManagerState.hpp"
 
 namespace cppmicroservices {
@@ -45,7 +49,7 @@ public:
   CMDisabledState();
 
   /**
-   * Contruct with a future object
+   * Construct with a future object
    *
    * /param fut is the future associated with this state. This future represents
    *        the task performed when the ComponentManager changes it's state from