+++ install_valgrind.sh
@@ -1,8 +1,8 @@
 #! /usr/bin/env bash
 #
-wget https://sourceware.org/pub/valgrind/valgrind-3.15.0.tar.bz2
-tar -xjf valgrind-3.15.0.tar.bz2
-pushd valgrind-3.15.0
+wget https://sourceware.org/pub/valgrind/valgrind-3.16.1.tar.bz2
+tar -xjf valgrind-3.16.1.tar.bz2
+pushd valgrind-3.16.1
 ./configure --prefix=/usr/local
 make
 sudo make install